+++ src/views/ReviewSpot.vue
@@ -168,10 +168,6 @@
                         <label for="uploadImages">Upload Images:</label>
                         <ImageUpload
                             v-model:images="SO.uploadImages"
-                            :max-image-count="4"
-                            :ID="SO.spotId"
-                            v-model:startUpload="uploadImages"
-                            @UploadImageURLs="uploadedImageURLS"
                         ></ImageUpload>
                     </div>
                 </div>
@@ -415,7 +411,6 @@ export default {
                 title: '',
             },
             initialFormData: {},
-            uploadImages: false,
         };
     },
     computed: {
@@ -553,7 +548,6 @@ export default {
                     }
                 }
             });
-            this.uploadImages = true;
             this.setUpdatedFields(Array.from(updatedFields));
             this.saveForm().then((response) => {
                 if (!response.DisplayMsg) this.updateInitialFormState();
